Dv 16 apply the job tracer and tracker where its relevant in the devdox ai agent (#19)
* refactor: Added a retry mechanism alongside orchestration and tracking while the process goes on

Description:
Added the logic responsible for tracking and tracing the job while its performed

* refactor: Added a retry mechanism alongside orchestration and tracking while the process goes on

Description:
Implemented the tracker and tracer into the code alongside the retry mechanism to complement it.

* refactor: Added a retry mechanism alongside orchestration and tracking while the process goes on

Description:
Bumped up the version of devddox-ai-models package
added `; sys_platform != 'win32'` to safeguard against downloading uvloop dependency on windows, since its not supported. this will automate the process of deciding whether to install it or not, rather than having to constantly comment and uncomment it, moreover, it makes its clear from the get go that it does not work with windows

* replaced the Job Level Analysis with a more meaningful levels:
- LOAD_TESTS = "loading_tests"

- SWAGGER_TEST_GENERATION="swagger_test_generation"

---------

Co-authored-by: Mohamad Ali Jaafar <mohamadali.jaafar@montyholding.com>