Fix: Gmail keeps picking plain text by default
Description:
Unlike outlook which is less strict, where it automatically picks the HTML. Gmail on the other hand, is very strict, and depending on how you set up Multipart Subtype it will pick the plain text automatically.

Solution:
THe solution to eliminate this headache is to eliminate the MultipartSubtypeEnum manual setting, and let FastAPI mail handle it automatically, thus solving the issue.