Dv 38 investigate and fix gmails undesired handling of multipart emails with html and plain text alternatives (#29)
* enhancement: enhanced the readability of the code by adding proper types

* Fix: Bumped up the version of FastAPI mail

AS per the recommendation of FastAPI devs, the pydantic issue has been solved in version `1.5.1`. I have bumped up the version of Fast API mail to 1.5.1 which introduced the fix, thus we do not need the temporary solution `temporary_fastapi_message_schema.py` anymore

* Fix: Gmail keeps picking plain text by default

Description:
Unlike outlook which is less strict, where it automatically picks the HTML. Gmail on the other hand, is very strict, and depending on how you set up Multipart Subtype it will pick the plain text automatically.

Solution:
THe solution to eliminate this headache is to eliminate the MultipartSubtypeEnum manual setting, and let FastAPI mail handle it automatically, thus solving the issue.

---------

Co-authored-by: Mohamad Ali Jaafar <mohamadali.jaafar@montyholding.com>